BUG: The percentiles need to be computed using a different method
The default method will be to compute the minimum and maximum intensity
in the image or labels. Generate the sample and then calculate the
percentiles in the distribution. Once the limits are calculated, a
histogram is generated with out clipping the measurements outside of
the range.

+++ src/lib/itkTextureFilter/itkScalarImageToConnectedIntensitySizeListSampleFilter.h
@@ -94,11 +94,11 @@ public:
     itkSetMacro(NumberOfIntensityBins, int)
     itkGetMacro(NumberOfIntensityBins, int)
 
-    itkSetMacro(MinSize, int)
-    itkGetMacro(MinSize, int)
+    itkSetMacro(MinSize, double)
+    itkGetMacro(MinSize, double)
 
-    itkSetMacro(MaxSize, int)
-    itkGetMacro(MaxSize, int)
+    itkSetMacro(MaxSize, double)
+    itkGetMacro(MaxSize, double)
 
     itkSetMacro(BackgroundValue, InputImagePixelType)
     itkGetMacro(BackgroundValue, InputImagePixelType)
@@ -130,8 +130,8 @@ private:
   InputImagePixelType m_BackgroundValue;
   int m_NumberOfIntensityBins;
 
-  int m_MinSize;
-  int m_MaxSize;
+  double m_MinSize;
+  double m_MaxSize;
   int m_IntensityBinSize;
 
   bool m_FullConnectivity;
